Make pre-existing boolean 'booleanDeprecated'
