+++ prismic-model/js/landing-pages.js
@@ -6,7 +6,7 @@ import link from './parts/link';
 import timestamp from './parts/timestamp';
 import structuredText from './parts/structured-text';
 import captionedImageSlice from './parts/captioned-image-slice';
-import boolean from './parts/boolean';
+import booleanDeprecated from './parts/boolean-deprecated';
 
 const LandingPage = {
   Page: {
@@ -26,7 +26,7 @@ const LandingPage = {
           contentList: slice('Content list', {
             nonRepeat: {
               title,
-              hasFeatured: boolean('Feature the first item?'),
+              hasFeatured: booleanDeprecated('Feature the first item?'),
             },
             repeat: {
               content: link('Content item', 'document', [
