+++ prismic-model/js/parts/boolean-deprecated.js
@@ -1,0 +1,6 @@
+// @flow
+import select from './select';
+
+export default function(label: string) {
+  return select(label, ['yes']);
+}
