--- prismic-model/js/parts/boolean-new.js
@@ -1,11 +1,0 @@
-// @flow
-
-export default function(label: string, defaultValue: boolean = false) {
-  return {
-    type: 'Boolean',
-    config: {
-      default_value: defaultValue,
-      label: label,
-    },
-  };
-}
