+++ prismic-model/js/popup-dialog.js
@@ -2,7 +2,7 @@
 
 import structuredText from './parts/structured-text';
 import text from './parts/text';
-import booleanNew from './parts/boolean-new';
+import boolean from './parts/boolean';
 import link from './parts/link';
 
 const PopupDialog = {
@@ -12,7 +12,7 @@ const PopupDialog = {
     dialogCopy: structuredText('Text inside the open dialog', 'single'),
     ctaText: text('CTA inside the open dialog button text'),
     ctaLink: link('CTA inside the open dialog button link', 'web'),
-    isShown: booleanNew('Is shown?', false),
+    isShown: boolean('Is shown?', false),
   },
 };
